Separate notebook extension test from NotebookInstallSpec
